requirements.txt
Pillow>=10.0.0
gitpython>=3.1.0
matplotlib>=3.5.0